backend/requirements.txt
fastapi[standard]==0.112.0
tortoise-orm
passlib[bcrypt]
pydantic>=2.0.0
python-dotenv
pyjwt
pytest
httpx